Crate fann

Structs

CascadeParams
Parameters for cascade training.

Fann

FannError

TrainData

Enums

ActivationFunc
The activation functions used for the neurons during training. They can either be set for a group of neurons using set_activation_func_hidden and set_activation_func_output, or for a single neuron using set_activation_func.

ErrorFunc
Error function used during training.

FannErrorType

NetType
Network types

StopFunc
Stop critieria for training.

TrainAlgorithm
The Training algorithms used when training on fann_train_data with functions like fann_train_on_data or fann_train_on_file. The incremental training alters the weights after each time it is presented an input pattern, while batch only alters the weights once after it has been presented to all the patterns.

Type Definitions

Connection

FannResult