Struct fann::CascadeParams

pub struct CascadeParams {
    pub output_change_fraction: c_float,
    pub output_stagnation_epochs: c_uint,
    pub candidate_change_fraction: c_float,
    pub candidate_stagnation_epochs: c_uint,
    pub candidate_limit: fann_type,
    pub weight_multiplier: fann_type,
    pub max_out_epochs: c_uint,
    pub max_cand_epochs: c_uint,
    pub activation_functions: Vec<ActivationFunc>,
    pub activation_steepnesses: Vec<fann_type>,
    pub num_candidate_groups: c_uint,
}

Parameters for cascade training.

Fields

output_change_fraction

A number between 0 and 1 determining how large a fraction the mean square error should change within output_stagnation_epochs during training of the output connections, in order for the training to stagnate. After stagnation, training of the output connections ends and new candidates are prepared.

This means: If the MSE does not change by a fraction of output_change_fraction during a period of output_stagnation_epochs, the training of the output connections is stopped because training has stagnated.

output_stagnation_epochs

The number of epochs training is allowed to continue without changing the MSE by a fraction of at least output_change_fraction.

candidate_change_fraction

A number between 0 and 1 determining how large a fraction the mean square error should change within candidate_stagnation_epochs during training of the candidate neurons, in order for the training to stagnate. After stagnation, training of the candidate neurons is stopped and the best candidate is selected.

This means: If the MSE does not change by a fraction of candidate_change_fraction during a period of candidate_stagnation_epochs, the training of the candidate neurons is stopped because training has stagnated.

candidate_stagnation_epochs

The number of epochs training is allowed to continue without changing the MSE by a fraction of candidate_change_fraction.

candidate_limit

A limit for how much the candidate neuron may be trained. It limits the ratio between the MSE and the candidate score.

weight_multiplier

Multiplier for the weight of the candidate neuron before adding it to the network. Usually between 0 and 1, to make training less aggressive.

max_out_epochs

The maximum number of epochs the output connections may be trained after adding a new candidate neuron.

max_cand_epochs

The maximum number of epochs the input connections to the candidates may be trained before adding a new candidate neuron.

activation_functions

The activation functions for the candidate neurons.

activation_steepnesses

The activation function steepness values for the candidate neurons.

num_candidate_groups

The number of candidate neurons to be trained for each combination of activation function and steepness.

Methods

impl CascadeParams

fn get_num_candidates(&self) -> c_uint

The number of candidates used during training: the number of combinations of activation functions and steepnesses, times num_candidate_groups.

For every combination of activation function and steepness, num_candidate_groups such neurons, with different initial weights, are trained.