Struct fann::Fann

pub struct Fann {
    // some fields omitted
}

Methods

impl Fann

fn new(layers: &[c_uint]) -> FannResult<Fann>

Create a fully connected neural network.

There will be a bias neuron in each layer except the output layer, and this bias neuron will be connected to all neurons in the next layer. When running the network, the bias nodes always emit 1.

Arguments

layers - Specifies the number of neurons in each layer, starting with the input and ending with the output layer.

Example

// Creating a network with 2 input neurons, 1 output neuron,
// and two hidden layers with 8 and 9 neurons.
let layers = [2, 8, 9, 1];
fann::Fann::new(&layers).unwrap();

fn new_sparse(connection_rate: c_float, layers: &[c_uint]) -> FannResult<Fann>

Create a neural network that is not necessarily fully connected.

There will be a bias neuron in each layer except the output layer, and this bias neuron will be connected to all neurons in the next layer. When running the network, the bias nodes always emit 1.

Arguments

connection_rate - The share of pairs of neurons in consecutive layers that will be connected.

layers - Specifies the number of neurons in each layer, starting with the input and ending with the output layer.

fn new_shortcut(layers: &[c_uint]) -> FannResult<Fann>

Create a neural network which has shortcut connections, i. e. it doesn't connect only each layer to its successor, but every layer with every later layer: Each neuron has connections to all neurons in all subsequent layers.

fn train(&mut self, input: &[fann_type], desired_output: &[fann_type]) -> FannResult<()>

Train with a single pair of input and output. This is always incremental training (see TrainAlg), since only one pattern is presented.

fn train_on_data(&mut self, data: &TrainData, max_epochs: c_uint, epochs_between_reports: c_uint, desired_error: c_float) -> FannResult<()>

Train the network on the given data set.

Arguments

data - The training data.

max_epochs - The maximum number of training epochs.

epochs_between_reports - The number of epochs between printing a status report to stdout, or 0 to print no reports.

desired_error - The desired maximum value of get_mse or get_bit_fail, depending on which stop function was selected.

fn train_on_file<P: AsRef<Path>>(&mut self, path: P, max_epochs: c_uint, epochs_between_reports: c_uint, desired_error: c_float) -> FannResult<()>

Do the same as train_on_data but read the training data directly from a file.

fn train_epoch(&mut self, data: &TrainData) -> FannResult<c_float>

Train one epoch with a set of training data, i. e. each sample from the training data is considered exactly once.

Returns the mean square error as it is calculated either before or during the actual training. This is not the actual MSE after the training epoch, but since calculating this will require to go through the entire training set once more, it is more than adequate to use this value during training.

fn test(&mut self, input: &[fann_type], desired_output: &[fann_type]) -> FannResult<Vec<fann_type>>

Test with a single pair of input and output. This operation updates the mean square error but does not change the network.

Returns the actual output of the network.

fn test_data(&mut self, data: &TrainData) -> FannResult<c_float>

Test with a training data set and calculate the mean square error.

fn get_mse(&self) -> c_float

Get the mean square error.

fn get_bit_fail(&self) -> c_uint

Get the number of fail bits, i. e. the number of neurons which differed from the desired output by more than the bit fail limit since the previous reset.

fn reset_mse_and_bit_fail(&mut self)

Reset the mean square error and bit fail count.

fn run(&self, input: &[fann_type]) -> FannResult<Vec<fann_type>>

Run the input through the neural network and returns the output. The length of the input must equal the number of input neurons and the length of the output will equal the number of output neurons.

fn get_num_input(&self) -> c_uint

Get the number of input neurons.

fn get_num_output(&self) -> c_uint

Get the number of output neurons.

fn get_total_neurons(&self) -> c_uint

Get the total number of neurons, including the bias neurons.

E. g. a 2-4-2 network has 3 + 5 + 2 = 10 neurons (because two layers have bias neurons).

fn get_total_connections(&self) -> c_uint

Get the total number of connections.

fn get_network_type(&self) -> NetType

Get the type of the neural network.

fn get_connection_rate(&self) -> c_float

Get the connection rate used when the network was created.

fn get_num_layers(&self) -> c_uint

Get the number of layers in the network.

fn get_activation_func(&self, layer: c_int, neuron: c_int) -> FannResult<ActivationFunc>

Get the activation function for neuron number neuron in layer number layer, counting the input layer as number 0. Input layer neurons do not have an activation function, so layer must be at least 1.

fn set_activation_func(&mut self, af: ActivationFunc, layer: c_int, neuron: c_int)

Set the activation function for neuron number neuron in layer number layer, counting the input layer as number 0. Input layer neurons do not have an activation function, so layer must be at least 1.

fn set_activation_func_hidden(&mut self, activation_func: ActivationFunc)

Set the activation function for all hidden layers.

fn set_activation_func_output(&mut self, activation_func: ActivationFunc)

Set the activation function for the output layer.

fn get_activation_steepness(&self, layer: c_int, neuron: c_int) -> Option<fann_type>

Get the activation steepness for neuron number neuron in layer number layer.

fn set_activation_steepness(&self, steepness: fann_type, layer: c_int, neuron: c_int)

Set the activation steepness for neuron number neuron in layer number layer, counting the input layer as number 0. Input layer neurons do not have an activation steepness, so layer must be at least 1.

The steepness determines how fast the function goes from minimum to maximum. A higher value will result in more aggressive training.

A steep activation function is adequate if outputs are binary, e. e. they are supposed to be either almost 0 or almost 1.

The default value is 0.5.

fn set_activation_steepness_layer(&self, steepness: fann_type, layer: c_int)

Set the activation steepness for layer number layer.

fn set_activation_steepness_hidden(&self, steepness: fann_type)

Set the activation steepness for all hidden layers.

fn set_activation_steepness_output(&self, steepness: fann_type)

Set the activation steepness for the output layer.

fn get_error_func(&self) -> ErrorFunc

Get the error function used during training.

fn set_error_func(&mut self, ef: ErrorFunc)

Set the error function used during training.

The default is Tanh.

fn get_stop_func(&self) -> StopFunc

Get the stop criterion for training.

fn set_stop_func(&mut self, sf: StopFunc)

Set the stop criterion for training.

The default is Mse.

fn get_bit_fail_limit(&self) -> fann_type

Get the bit fail limit.

fn set_bit_fail_limit(&mut self, bit_fail_limit: fann_type)

Set the bit fail limit.

Each output neuron value that differs from the desired output by more than the bit fail limit is counted as a failed bit.

fn get_train_algorithm(&self) -> TrainAlgorithm

Get the currently configured training algorithm.

fn set_train_algorithm(&mut self, ta: TrainAlgorithm)

Set the algorithm to be used for training.

fn get_learning_rate(&self) -> c_float

Get the learning rate, which is used to determine how aggressive training should be (not used by the RPROP algorithm). The default is 0.7.

fn set_learning_rate(&mut self, learning_rate: c_float)

Set the learning rate, which is used to determine how aggressive training should be (not used by the RPROP algorithm). The default is 0.7.

fn get_learning_momentum(&self) -> c_float

Get the learning momentum used in incremental training. It is recommended to use a value between 0.0 and 1.0. The default is 1.0.

fn set_learning_momentum(&mut self, learning_momentum: c_float)

Set the learning momentum used in incremental training. It is recommended to use a value between 0.0 and 1.0. The default is 1.0.